Fix issues with wave functions and frequencies

+++ src/lib/core/composition_state.rs
@@ -9,12 +9,12 @@ pub struct CompositionState {
     /// How far through, in steps, we are through the composition
     pub tick: u64,
     /// The frequency of the composition
-    pub frequency: u64,
+    pub frequency: f32,
 }
 
 impl CompositionState {
     #[allow(missing_docs)]
-    pub fn initial(frequency: u64) -> Self {
+    pub fn initial(frequency: f32) -> Self {
         CompositionState {
             time: Duration::from_secs(0),
             tick: 0,
@@ -25,7 +25,8 @@ impl CompositionState {
     /// Step to the next state in the composition
     pub fn increment(&mut self) {
         self.tick += 1;
-        self.time =
-            Duration::from_nanos((self.tick * 1_000_000) / self.frequency);
+        self.time = Duration::from_nanos(
+            (self.tick * 1_000_000) / self.frequency as u64,
+        );
     }
 }
